requirements.txt
Flask
python-dotenv
Flask-SQLAlchemy
flask-login
Werkzeug
Flask-Mail
Flask-Migrate